Last commit

+++ src/contexts/CitiesContext.jsx
@@ -1,43 +1,99 @@
-import { createContext, useContext } from "react";
-import { useEffect, useState } from "react";
+import { createContext, useContext, useReducer } from "react";
+import { useEffect } from "react";
 
 const CitiesContext = createContext();
 
 const BaseUrl = "http://localhost:8000";
 
+const reducerInitialState = {
+  cities: [],
+  loading: false,
+  currentCity: {},
+  error: "",
+};
+
+function reducerFunction(state, action) {
+  switch (action.type) {
+    case "loading":
+      return {
+        ...state,
+        loading: true,
+      };
+    case "cities/loaded":
+      return {
+        ...state,
+        loading: false,
+        cities: action.payload,
+      };
+    case "city/loaded":
+      return {
+        ...state,
+        loading: false,
+        currentCity: action.payload,
+      };
+    case "city/created":
+      return {
+        ...state,
+        loading: false,
+        currentCity: action.payload,
+        cities: [...state.cities, action.payload],
+      };
+    case "city/deleted":
+      return {
+        ...state,
+        loading: false,
+        cities: state.cities.filter((city) => city != action.payload),
+        currentCity: {},
+      };
+    case "rejected":
+      return {
+        ...state,
+        loading: false,
+        error: action.payload,
+      };
+    default:
+      return state;
+  }
+}
+
 // eslint-disable-next-line react/prop-types
 function CitiesProvider({ children }) {
-  const [cities, setCities] = useState([]);
-  const [loading, setLoading] = useState(false);
-  const [currentCity, setCurrentCity] = useState({});
+  const [{ cities, loading, currentCity, error }, dispatcher] = useReducer(
+    reducerFunction,
+    reducerInitialState
+  );
 
   useEffect(function () {
-    setLoading(true);
+    dispatcher({ type: "loading" });
     const fetchCities = async function () {
       try {
         (await fetch(`${BaseUrl}/cities`)).json().then((data) => {
-          setCities(data);
+          dispatcher({ type: "cities/loaded", payload: data });
         });
       } catch (error) {
         alert("There is an error in fetching cities", error);
-      } finally {
-        setLoading(false);
+        dispatcher({
+          type: "rejected",
+          payload: "There is an error in fetching cities..." + error,
+        });
       }
     };
     fetchCities();
   }, []);
 
   function getCity(id) {
-    setLoading(true);
+    dispatcher({ type: "loading" });
     const fetchCities = async function () {
       try {
         (await fetch(`${BaseUrl}/cities/${id}`)).json().then((data) => {
-          setCurrentCity(data);
+          dispatcher({ type: "city/loaded", payload: data });
         });
       } catch (error) {
         alert("There is an error in fetching cities", error);
-      } finally {
-        setLoading(false);
+        dispatcher({
+          type: "rejected",
+          payload: "There is an error in fetching city..." + error,
+        });
       }
     };
     fetchCities();
@@ -45,7 +101,7 @@ function CitiesProvider({ children }) {
 
   async function createCity(city) {
     try {
-      setLoading(true);
+      dispatcher({ type: "loading" });
       const res = await fetch(`${BaseUrl}/cities`, {
         method: "POST",
         body: JSON.stringify(city),
@@ -54,27 +110,37 @@ function CitiesProvider({ children }) {
         },
       });
       const data = await res.json();
-      setCurrentCity(data);
-      setCities((cities) => [...cities, data]);
+      dispatcher({
+        type: "cities/created",
+        payload: data,
+      });
     } catch (error) {
       alert("There was sone error loading the data...");
-    } finally {
-      setLoading(false);
+      dispatcher({
+        type: "rejected",
+        payload: "There is an error in creating new city..." + error,
+      });
     }
   }
 
   async function deleteCity(id) {
+    if (currentCity.id === id) return;
     try {
-      setLoading(true);
+      dispatcher({ type: "loading" });
       const res = await fetch(`${BaseUrl}/cities/${id}`, {
         method: "DELETE",
       });
       if (res.status === 200)
-        setCities((cities) => cities.filter((city) => city != id));
+        dispatcher({
+          type: "city/deleted",
+          payload: id,
+        });
     } catch (error) {
       alert("There was sone error deleting the city...");
-    } finally {
-      setLoading(false);
+      dispatcher({
+        type: "rejected",
+        payload: "There is an error in deleting city..." + error,
+      });
     }
   }
   return (
@@ -84,6 +150,7 @@ function CitiesProvider({ children }) {
         loading,
         currentCity,
         getCity,
+        error,
         createCity,
         deleteCity,
       }}
